move basic gengou type classes on japan module
