+++ spec/zakuro/calculation/era/gengou/dated_scroll_spec.rb
@@ -53,7 +53,7 @@ describe 'Zakuro' do
               list.instance_variable_set(
                 '@list', [
                   Zakuro::Japan::Gengou::Alignment::LinearGengou.new(
-                    gengou: Zakuro::Japan::Gengou::Resource::Gengou.new(
+                    gengou: Zakuro::Japan::Type::Base::Gengou.new(
                       name: '元号',
                       start_year: Zakuro::Japan::Type::Base::Both::Year.new(
                         japan: 1,
@@ -140,7 +140,7 @@ describe 'Zakuro' do
               list.instance_variable_set(
                 '@list', [
                   Zakuro::Japan::Gengou::Alignment::LinearGengou.new(
-                    gengou: Zakuro::Japan::Gengou::Resource::Gengou.new(
+                    gengou: Zakuro::Japan::Type::Base::Gengou.new(
                       name: '元号',
                       start_year: Zakuro::Japan::Type::Base::Both::Year.new(
                         japan: 1,
@@ -240,7 +240,7 @@ describe 'Zakuro' do
               list.instance_variable_set(
                 '@list', [
                   Zakuro::Japan::Gengou::Alignment::LinearGengou.new(
-                    gengou: Zakuro::Japan::Gengou::Resource::Gengou.new(
+                    gengou: Zakuro::Japan::Type::Base::Gengou.new(
                       name: '元号',
                       start_year: Zakuro::Japan::Type::Base::Both::Year.new(
                         japan: 1,
@@ -357,7 +357,7 @@ describe 'Zakuro' do
               list.instance_variable_set(
                 '@list', [
                   Zakuro::Japan::Gengou::Alignment::LinearGengou.new(
-                    gengou: Zakuro::Japan::Gengou::Resource::Gengou.new(
+                    gengou: Zakuro::Japan::Type::Base::Gengou.new(
                       name: '元号',
                       start_year: Zakuro::Japan::Type::Base::Both::Year.new(
                         japan: 1,
