+++ spec/zakuro/calculation/era/gengou/named_scroll_spec.rb
@@ -66,7 +66,7 @@ describe 'Zakuro' do
               list.instance_variable_set(
                 '@list', [
                   Zakuro::Japan::Gengou::Alignment::LinearGengou.new(
-                    gengou: Zakuro::Japan::Gengou::Resource::Gengou.new(
+                    gengou: Zakuro::Japan::Type::Base::Gengou.new(
                       name: '元号',
                       start_year: Zakuro::Japan::Type::Base::Both::Year.new(
                         japan: 1,
