+++ spec/zakuro/era/japan/gengou/resource/parser_spec.rb
@@ -42,7 +42,7 @@ describe 'Zakuro' do
               end
               it 'is created Set class' do
                 actual = Zakuro::Japan::Gengou::Resource::Parser.run(filepath: TEST_YAML_PATH)
-                expect(actual).to be_a(Zakuro::Japan::Gengou::Resource::Set)
+                expect(actual).to be_a(Zakuro::Japan::Type::Base::GengouSet)
               end
             end
             context 'internal data' do
@@ -81,7 +81,7 @@ describe 'Zakuro' do
               context 'list item data type' do
                 it 'is created gengou class' do
                   item = actual.list[0]
-                  expect(item).to be_a(Zakuro::Japan::Gengou::Resource::Gengou)
+                  expect(item).to be_a(Zakuro::Japan::Type::Base::Gengou)
                 end
               end
               context 'first-list[0]' do
